Add reserve(), checkIn() methods.
Idea is that reserve() allocates space and returns a pointer to the new
object. checkIn() later advances the tail-pointer. This should give the
"producer" thread access to the new object. This should reduce the
amount of data which needs to be copied. Example would be a list of
audio buffers.